Refactor check_ssl
